Moves all "User" related classes into a single namespace
This approach keeps all related classes together, making it easier to
find which classes go with what.
