+++ src/App/RestDispatchTrait.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Handler;
+namespace App;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
